requirements.txt
asgiref==3.8.1
Brotli==1.1.0
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
cssselect2==0.7.0
Django==5.0.6
django-phonenumber-field==8.0.0
django-tables2==2.7.0
django-widget-tweaks==1.5.0
fonttools==4.53.1
html5lib==1.1
idna==3.7
pillow==10.4.0
pycparser==2.22
pydyf==0.11.0
PyJWT==2.8.0
pyphen==0.16.0
requests==2.32.3
six==1.16.0
sqlparse==0.5.1
tinycss2==1.3.0
urllib3==2.2.2
weasyprint==62.3
webencodings==0.5.1
zopfli==0.2.3
